+++ front/src/component/admin/cocktail/InputContainer.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+const InputContainer = ({property, value, onChange, onEnter}) => {
+    return (
+        <div className="inputContainer">
+            <div>{property}</div>
+            <input
+                className="inputBox"
+                name={property}
+                onChange={onChange}
+                value={value}
+                onKeyDown={onEnter}
+            />
+        </div>
+    );
+};
+
+export default InputContainer;
